pyproject.toml
[project]
name = "2025-fastapi-seminar-template-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "authlib>=1.6.4",
    "fastapi[standard]>=0.117.1",
    "pydantic-settings>=2.10.1",
    "pymysql>=1.1.2",
    "pytest>=8.4.2",
    "sqlalchemy>=2.0.43",
]
